Refact: City 모델 파일 분리
