+++ WeatherForecast/WeatherForecast/Models/City.swift
@@ -1,0 +1,17 @@
+//
+//  City.swift
+//  WeatherForecast
+//
+//  Created by JunHeeJo on 2/8/24.
+//
+
+import Foundation
+
+final class City: Decodable {
+    let id: Int
+    let name: String
+    let coord: Coord
+    let country: String
+    let population, timezone: Int
+    let sunrise, sunset: TimeInterval
+}
